ReReRe add login

+++ app/Http/Controllers/TasksController.php
@@ -81,10 +81,18 @@ class TasksController extends Controller
         $task = Task::find($id);
         
         if (\Auth::check()) {
-            return view(
-                'tasks.show',
-                ['task' => $task]
+            
+            if($task->user_id == \Auth::user()->id) {
+                
+                return view(
+                    'tasks.show',
+                    ['task' => $task]
                 );
+                
+            }else{
+                return redirect('/'); 
+            }
+            
         }else {
             return redirect('/'); 
         }
@@ -100,17 +108,22 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
-         if (\Auth::check()) {
-            return view(
-                'tasks.edit',
-                ['task' => $task]
+        if (\Auth::check()) {
+            if ($task->user_id == \Auth::user()->id){
+                
+                return view(
+                    'tasks.edit',
+                    ['task' => $task]
                 );
-        }else {
+                
+            }else{
+                return redirect('/');
+                
+            }
+            
+        }else{
             return redirect('/'); 
-        }
-        // return view('tasks.edit', [
-        //     'task' => $task,
-        //     ]);
+        } 
     }
 
     /**
@@ -122,17 +135,30 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'status' => 'required|max:10',
-            'content' => 'required|max:191',
-        ]);
-        
-        $task = Task::find($id);
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-        
-        return redirect('/');
+        if(\Auth::check()) {
+            
+            $task = Task::find($id);
+            
+            if($task->user_id == \Auth::user()->id) {
+                $this->validate($request, [
+                    'status' => 'required|max:10',
+                    'content' => 'required|max:191',
+                    ]);
+                    
+                // $task = Task::find($id);
+                $task->status = $request->status;
+                $task->content = $request->content;
+                $task->save();
+                    
+                return redirect('/');
+                
+            }else {
+                return redirect('/');
+            }
+            
+        }else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -143,9 +169,22 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
-        
-        return redirect('/');
+        if (\Auth::check()) {
+            
+            $task = Task::find($id);
+            
+            if($task->user_id == \Auth::user()->id) {
+                // $task = Task::find($id);
+                $task->delete();
+                
+                return redirect('/');
+                
+            }else {
+                return redirect('/');
+            }
+            
+        }else {
+            return redirect('/');
+        }
     }
 }
